update commands to use v13 methods

+++ src/commands/moderation/purge.ts
@@ -22,15 +22,17 @@ export default class PurgeCommand extends Command {
   }
 
   public async main({ msg, args }: RunArgumentsOptions) {
-    const amount = Math.min(10000, Math.max(0, parseInt(args.shift()) + 1));
+    if (msg.channel.type !== 'GUILD_TEXT') return;
+    const amount = Math.min(10000, Math.max(0, parseInt(args.shift(), 10) + 1));
     let user: User = null;
     let messages: Collection<string, Message>;
     if (args[0]) {
       user = await this.client.util.getUser(msg, args.join(' '));
-      messages = (await msg.channel.messages.fetch({ limit: amount }, false)).filter((m) => m.author.id === user.id);
+      messages = (await msg.channel.messages
+        .fetch({ limit: amount })).filter((m) => m.author.id === user.id);
     }
 
     const { size } = await msg.channel.bulkDelete(user ? messages : amount, true);
-    msg.channel.send(`Deleted \`${size}\` messages.\nRemember bulk only deletes messages younger than 2 weeks.`);
+    return `Deleted \`${size}\` messages.\nRemember bulk only deletes messages younger than 2 weeks.`;
   }
 }
